Reorganize Playground CLI code to exclude unused code
