+++ packages/playground/cli/src/cli.ts
@@ -1,151 +1,3 @@
-/* eslint-disable no-console */
-import { SupportedPHPVersions } from '@php-wasm/universal';
-import { RecommendedPHPVersion } from '@wp-playground/common';
-import yargs from 'yargs';
-import {
-	parseMountDirArguments,
-	parseMountWithDelimiterArguments,
-} from './mount';
-import type { RunCLIArgs } from './run-cli';
-import { runCLI } from './run-cli';
+import { parseOptionsAndRunCLI } from './run-cli';
 
-async function run() {
-	/**
-	 * @TODO This looks similar to Query API args https://wordpress.github.io/wordpress-playground/developers/apis/query-api/
-	 *       Perhaps the two could be handled by the same code?
-	 */
-	const yargsObject = yargs(process.argv.slice(2))
-		.usage('Usage: wp-playground <command> [options]')
-		.positional('command', {
-			describe: 'Command to run',
-			choices: ['server', 'run-blueprint', 'build-snapshot'] as const,
-			demandOption: true,
-		})
-		.option('outfile', {
-			describe: 'When building, write to this output file.',
-			type: 'string',
-			default: 'wordpress.zip',
-		})
-		.option('port', {
-			describe: 'Port to listen on when serving.',
-			type: 'number',
-			default: 9400,
-		})
-		.option('php', {
-			describe: 'PHP version to use.',
-			type: 'string',
-			default: RecommendedPHPVersion,
-			choices: SupportedPHPVersions,
-		})
-		.option('wp', {
-			describe: 'WordPress version to use.',
-			type: 'string',
-			default: 'latest',
-		})
-		// @TODO: Support read-only mounts, e.g. via WORKERFS, a custom
-		// ReadOnlyNODEFS, or by copying the files into MEMFS
-		.option('mount', {
-			describe:
-				'Mount a directory to the PHP runtime. You can provide --mount multiple times. Format: /host/path:/vfs/path',
-			type: 'array',
-			string: true,
-			coerce: parseMountWithDelimiterArguments,
-		})
-		.option('mountBeforeInstall', {
-			describe:
-				'Mount a directory to the PHP runtime before installing WordPress. You can provide --mount-before-install multiple times. Format: /host/path:/vfs/path',
-			type: 'array',
-			string: true,
-			coerce: parseMountWithDelimiterArguments,
-		})
-		.option('mountDir', {
-			describe:
-				'Mount a directory to the PHP runtime. You can provide --mount-dir multiple times. Format: "/host/path" "/vfs/path"',
-			type: 'array',
-			nargs: 2,
-			array: true,
-			// coerce: parseMountDirArguments,
-		})
-		.option('mountDirBeforeInstall', {
-			describe:
-				'Mount a directory to the PHP runtime before installing WordPress. You can provide --mount-before-install multiple times. Format: "/host/path" "/vfs/path"',
-			type: 'string',
-			nargs: 2,
-			array: true,
-			coerce: parseMountDirArguments,
-		})
-		.option('login', {
-			describe: 'Should log the user in',
-			type: 'boolean',
-			default: false,
-		})
-		.option('blueprint', {
-			describe: 'Blueprint to execute.',
-			type: 'string',
-		})
-		.option('blueprintMayReadAdjacentFiles', {
-			describe:
-				'Consent flag: Allow "bundled" resources in a local blueprint to read files in the same directory as the blueprint file.',
-			type: 'boolean',
-			default: false,
-		})
-		.option('skipWordPressSetup', {
-			describe:
-				'Do not download, unzip, and install WordPress. Useful for mounting a pre-configured WordPress directory at /wordpress.',
-			type: 'boolean',
-			default: false,
-		})
-		.option('skipSqliteSetup', {
-			describe:
-				'Skip the SQLite integration plugin setup to allow the WordPress site to use MySQL.',
-			type: 'boolean',
-			default: false,
-		})
-		.option('quiet', {
-			describe: 'Do not output logs and progress messages.',
-			type: 'boolean',
-			default: false,
-		})
-		.option('debug', {
-			describe:
-				'Print PHP error log content if an error occurs during Playground boot.',
-			type: 'boolean',
-			default: false,
-		})
-		.option('autoMount', {
-			describe: `Automatically mount the current working directory. You can mount a WordPress directory, a plugin directory, a theme directory, a wp-content directory, or any directory containing PHP and HTML files.`,
-			type: 'boolean',
-			default: false,
-		})
-		.option('followSymlinks', {
-			describe:
-				'Allow Playground to follow symlinks by automatically mounting symlinked directories and files encountered in mounted directories. \nWarning: Following symlinks will expose files outside mounted directories to Playground and could be a security risk.',
-			type: 'boolean',
-			default: false,
-		})
-		.showHelpOnFail(false);
-
-	yargsObject.wrap(yargsObject.terminalWidth());
-	const args = await yargsObject.argv;
-
-	const command = args._[0] as string;
-
-	if (!['run-blueprint', 'server', 'build-snapshot'].includes(command)) {
-		yargsObject.showHelp();
-		process.exit(1);
-	}
-
-	const cliArgs = {
-		...args,
-		command,
-		mount: [...(args.mount || []), ...(args.mountDir || [])],
-		mountBeforeInstall: [
-			...(args.mountBeforeInstall || []),
-			...(args.mountDirBeforeInstall || []),
-		],
-	} as RunCLIArgs;
-
-	return await runCLI(cliArgs);
-}
-
-await run();
+await parseOptionsAndRunCLI();
